move constants
