--- renderer/constants/routes.json
@@ -1,3 +1,0 @@
-{
-  "HOME": "/"
-}
